+++ com.microej.example.iot.ssl.rest/src/main/java/com/microej/example/iot/ssl/rest/ui/ConnectingPage.java
@@ -1,11 +1,17 @@
 package com.microej.example.iot.ssl.rest.ui;
 
-import ej.container.List;
 import ej.mwt.Widget;
 import ej.widget.basic.Label;
 import ej.widget.basic.drawing.CircularProgressBar;
+import ej.widget.container.List;
+import ej.widget.navigation.navigator.HistorizedNavigator;
 
 public class ConnectingPage extends AbstractOutputPage {
+
+	public ConnectingPage(HistorizedNavigator navi) {
+		super(navi);
+		// TODO Auto-generated constructor stub
+	}
 
 	@Override
 	protected Widget createMainContent() {