+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 14.0.0 (2017-14-06)
+
+Features:
+	- Add new com.microej.example.iot.ssl.restget example
+	- Migrating from MicroEJ 4.0 to MicroEJ 4.1
+
 ## 13.0.0 (2016-12-29)
 Features:
   - Change org